+++ izzet/navmesh/navmesh.go
@@ -1,20 +1,10 @@
 package navmesh
 
 import (
-	"container/heap"
-	"fmt"
-	"sort"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/go-gl/mathgl/mgl32"
 	"github.com/go-gl/mathgl/mgl64"
 	"github.com/kkevinchou/izzet/izzet/entities"
-	"github.com/kkevinchou/kitolib/collision"
 	"github.com/kkevinchou/kitolib/collision/collider"
 	"github.com/kkevinchou/kitolib/spatialpartition"
-	"github.com/kkevinchou/kitolib/utils"
 )
 
 // GENERAL IMPLEMENTATION NOTES
@@ -57,277 +47,6 @@ func New(world World) *NavigationMesh {
 	entity := world.GetEntityByID(3)
 	entities.SetLocalPosition(entity, mgl64.Vec3{0, -1000, 0})
 	return nm
-}
-
-func (n *NavigationMesh) voxelize() [][][]Voxel {
-	start := time.Now()
-	spatialPartition := n.world.SpatialPartition()
-	sEntities := spatialPartition.QueryEntities(n.Volume)
-
-	var candidateEntities []*entities.Entity
-	boundingBoxes := map[int]collider.BoundingBox{}
-	meshTriangles := map[int][]Triangle{}
-	entityTriCount := map[int]int{}
-
-	// collect candidate entities for voxelization
-	for _, entity := range sEntities {
-		e := n.world.GetEntityByID(entity.GetID())
-		if e == nil {
-			continue
-		}
-
-		if !strings.Contains(e.Model.Name(), "Tile") && !strings.Contains(e.Model.Name(), "Stair") {
-			continue
-		}
-
-		candidateEntities = append(candidateEntities, e)
-		boundingBoxes[e.GetID()] = *e.BoundingBox()
-		transform := utils.Mat4F64ToF32(entities.WorldTransform(e))
-
-		for _, rd := range e.Model.RenderData() {
-			meshID := rd.MeshID
-			mesh := e.Model.Collection().Meshes[meshID]
-			for i := 0; i < len(mesh.Vertices); i += 3 {
-				meshTriangles[meshID] = append(meshTriangles[meshID],
-					Triangle{
-						V1: utils.Vec3F32ToF64(transform.Mul4x1(mesh.Vertices[i].Position.Vec4(1)).Vec3()),
-						V2: utils.Vec3F32ToF64(transform.Mul4x1(mesh.Vertices[i+1].Position.Vec4(1)).Vec3()),
-						V3: utils.Vec3F32ToF64(transform.Mul4x1(mesh.Vertices[i+2].Position.Vec4(1)).Vec3()),
-					},
-				)
-			}
-			numVerts := len(mesh.Vertices)
-			entityTriCount[e.GetID()] = numVerts / 3
-		}
-	}
-
-	outputWork := make(chan OutputWork)
-	delta := n.Volume.MaxVertex.Sub(n.Volume.MinVertex)
-	var dimensions [3]int = [3]int{int(delta[0] / n.voxelDimension), int(delta[1] / n.voxelDimension), int(delta[2] / n.voxelDimension)}
-
-	inputWorkCount := dimensions[0] * dimensions[1] * dimensions[2]
-	inputWork := make(chan [3]int, inputWorkCount)
-	workerCount := 12
-
-	doneWorkerCount := 0
-	var doneWorkerMutex sync.Mutex
-
-	// set up workers perform voxelization at a specific 3d coordinate
-	for i := 0; i < workerCount; i++ {
-		go func() {
-			for input := range inputWork {
-				x, y, z := input[0], input[1], input[2]
-
-				voxel := &collider.BoundingBox{
-					MinVertex: n.Volume.MinVertex.Add(mgl64.Vec3{float64(x), float64(y), float64(z)}.Mul(n.voxelDimension)),
-					MaxVertex: n.Volume.MinVertex.Add(mgl64.Vec3{float64(x + 1), float64(y + 1), float64(z + 1)}.Mul(n.voxelDimension)),
-				}
-				voxelAABB := AABB{Min: voxel.MinVertex, Max: voxel.MaxVertex}
-
-				for _, entity := range candidateEntities {
-					bb := boundingBoxes[entity.GetID()]
-					if !collision.CheckOverlapAABBAABB(voxel, &bb) {
-						continue
-					}
-
-					for _, rd := range entity.Model.RenderData() {
-						for _, tri := range meshTriangles[rd.MeshID] {
-							if IntersectAABBTriangle(voxelAABB, tri) {
-								outputWork <- OutputWork{
-									x:           x,
-									y:           y,
-									z:           z,
-									boundingBox: *voxel,
-								}
-
-								goto Done
-							}
-						}
-					}
-				Done:
-				}
-			}
-
-			doneWorkerMutex.Lock()
-			doneWorkerCount++
-			if doneWorkerCount == workerCount {
-				fmt.Println("generation time seconds", time.Since(start).Seconds())
-				close(outputWork)
-			}
-			doneWorkerMutex.Unlock()
-		}()
-	}
-
-	// initialize the voxel field
-	voxelField := make([][][]Voxel, dimensions[0])
-	for i := range voxelField {
-		voxelField[i] = make([][]Voxel, dimensions[1])
-		for j := range voxelField[i] {
-			voxelField[i][j] = make([]Voxel, dimensions[2])
-		}
-	}
-
-	// create a work item for each voxel location
-	for i := 0; i < dimensions[0]; i++ {
-		for j := 0; j < dimensions[1]; j++ {
-			for k := 0; k < dimensions[2]; k++ {
-				inputWork <- [3]int{i, j, k}
-			}
-		}
-	}
-	close(inputWork)
-
-	// assemble voxels into the voxel field
-	for work := range outputWork {
-		n.voxelCount++
-
-		x, y, z := work.x, work.y, work.z
-		voxelField[x][y][z] = NewVoxel(x, y, z)
-		voxelField[x][y][z].Filled = true
-	}
-	fmt.Printf("generated %d voxels\n", n.voxelCount)
-
-	totalTricount := 0
-	for _, count := range entityTriCount {
-		totalTricount += count
-	}
-	fmt.Println("nav mesh entity tri count", totalTricount)
-	return voxelField
-}
-
-func buildNavigableArea(voxelField [][][]Voxel, dimensions [3]int) {
-	work := make(chan [2]int, dimensions[0]*dimensions[2])
-	workerCount := 12
-
-	var wg sync.WaitGroup
-	wg.Add(workerCount)
-
-	// remove voxels that have another voxel directly above it
-	for wc := 0; wc < workerCount; wc++ {
-		go func() {
-			defer wg.Done()
-			for w := range work {
-				for y := dimensions[1] - 1; y >= 0; y-- {
-					x, z := w[0], w[1]
-					if voxelField[x][y][z].Filled {
-						for i := 1; i < agentHeight+1; i++ {
-							if y-i < 0 {
-								break
-							}
-							voxelField[x][y-i][z] = NewVoxel(x, y-i, z)
-						}
-						y -= agentHeight
-					}
-				}
-			}
-		}()
-	}
-
-	for i := 0; i < dimensions[0]; i++ {
-		for j := 0; j < dimensions[2]; j++ {
-			work <- [2]int{i, j}
-		}
-	}
-
-	close(work)
-	wg.Wait()
-}
-
-func computeDistanceTransform(voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int) {
-	// boundaries have a distance field of 0
-	for y := 0; y < dimensions[1]; y++ {
-		for x := 0; x < dimensions[0]; x++ {
-			voxelField[x][y][0].DistanceField = 0
-			voxelField[x][y][dimensions[2]-1].DistanceField = 0
-		}
-		for z := 0; z < dimensions[2]; z++ {
-			voxelField[0][y][z].DistanceField = 0
-			voxelField[dimensions[0]-1][y][z].DistanceField = 0
-		}
-	}
-
-	for z := 1; z < dimensions[2]-1; z++ {
-		for x := 1; x < dimensions[0]-1; x++ {
-			for y := 0; y < dimensions[1]; y++ {
-				computeVoxelDistanceTransform(x, y, z, voxelField, reachField, dimensions, -1)
-			}
-		}
-	}
-
-	for z := dimensions[2] - 2; z > 0; z-- {
-		for x := dimensions[0] - 2; x > 0; x-- {
-			for y := 0; y < dimensions[1]; y++ {
-				computeVoxelDistanceTransform(x, y, z, voxelField, reachField, dimensions, 1)
-			}
-		}
-	}
-}
-
-func computeVoxelDistanceTransform(x, y, z int, voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int, sweepDir int) {
-	xNeighbor := voxelField[x+sweepDir][y][z]
-	zNeighbor := voxelField[x][y][z+sweepDir]
-	xzNegativeNeighbor := voxelField[x-1][y][z+sweepDir]
-	xzPositiveNeighbor := voxelField[x+1][y][z+sweepDir]
-
-	xReach := reachField[x+sweepDir][y][z]
-	zReach := reachField[x][y][z+sweepDir]
-	xzNegativeReach := reachField[x-1][y][z+sweepDir]
-	xzPositiveReach := reachField[x+1][y][z+sweepDir]
-
-	var minDistanceFieldValue float64 = MaxDistanceFieldValue
-
-	if (!xNeighbor.Filled && !xReach.hasSource) || (!zNeighbor.Filled && !zReach.hasSource) || (!xzNegativeNeighbor.Filled && !xzNegativeReach.hasSource) || (!xzPositiveNeighbor.Filled && !xzPositiveReach.hasSource) {
-		minDistanceFieldValue = 0
-	} else {
-		var distanceFieldXValue float64
-		if xNeighbor.Filled {
-			distanceFieldXValue = xNeighbor.DistanceField + 1
-		} else if xReach.hasSource {
-			source := xReach.source
-			distanceFieldXValue = voxelField[source[0]][source[1]][source[2]].DistanceField + 1
-		}
-
-		var distanceFieldZValue float64
-		if zNeighbor.Filled {
-			distanceFieldZValue = zNeighbor.DistanceField + 1
-		} else if zReach.hasSource {
-			source := zReach.source
-			distanceFieldZValue = voxelField[source[0]][source[1]][source[2]].DistanceField + 1
-		}
-
-		var distanceFieldXZNegativeValue float64
-		if xzNegativeNeighbor.Filled {
-			distanceFieldXZNegativeValue = xzNegativeNeighbor.DistanceField + 1.4
-		} else if xzNegativeReach.hasSource {
-			source := xzNegativeReach.source
-			distanceFieldXZNegativeValue = voxelField[source[0]][source[1]][source[2]].DistanceField + 1.4
-		}
-
-		var distanceFieldXZPositiveValue float64
-		if xzPositiveNeighbor.Filled {
-			distanceFieldXZPositiveValue = xzPositiveNeighbor.DistanceField + 1.4
-		} else if xzPositiveReach.hasSource {
-			source := xzPositiveReach.source
-			distanceFieldXZPositiveValue = voxelField[source[0]][source[1]][source[2]].DistanceField + 1.4
-		}
-
-		if distanceFieldXValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldXValue
-		}
-		if distanceFieldZValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldZValue
-		}
-		if distanceFieldXZNegativeValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldXZNegativeValue
-		}
-		if distanceFieldXZPositiveValue < minDistanceFieldValue {
-			minDistanceFieldValue = distanceFieldXZPositiveValue
-		}
-	}
-
-	if minDistanceFieldValue < voxelField[x][y][z].DistanceField {
-		voxelField[x][y][z].DistanceField = minDistanceFieldValue
-	}
 }
 
 type ReachInfo struct {
@@ -392,110 +111,6 @@ func neighborDist(voxel, neighbor *Voxel) float64 {
 	return dist
 }
 
-func watershed(voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int) map[int][][3]int {
-	pq := PriorityQueue{}
-
-	for x := 0; x < dimensions[0]; x++ {
-		for y := 0; y < dimensions[1]; y++ {
-			for z := 0; z < dimensions[2]; z++ {
-				if !voxelField[x][y][z].Filled {
-					continue
-				}
-
-				voxel := &voxelField[x][y][z]
-				pq = append(pq, &Item{value: voxel, priority: voxel.DistanceField})
-			}
-		}
-	}
-	heap.Init(&pq)
-
-	regionMap := map[int][][3]int{}
-	priorityMap := map[float64]int{}
-
-	// runCount := 0
-	for pq.Len() > 0 {
-		item := heap.Pop(&pq).(*Item)
-		voxel := item.value
-		priorityMap[item.priority]++
-
-		x := voxel.X
-		y := voxel.Y
-		z := voxel.Z
-
-		isSeed := true
-		var nearestNeighbor *Voxel
-
-		neighbors := getNeighbors(x, y, z, voxelField, reachField, dimensions)
-		for _, neighbor := range neighbors {
-			if neighbor.RegionID == -1 {
-				continue
-			}
-			if mgl32.FloatEqualThreshold(float32(voxel.DistanceField), float32(neighbor.DistanceField), 0.00000001) {
-				// voxel is from the same pass, don't consider neighbors that happened to be processed at the same time
-				continue
-			}
-
-			if nearestNeighbor == nil {
-				nearestNeighbor = neighbor
-			} else if neighborDist(voxel, neighbor) < neighborDist(voxel, nearestNeighbor) {
-				nearestNeighbor = neighbor
-			}
-			if neighbor.DistanceField > voxel.DistanceField {
-				isSeed = false
-			}
-		}
-
-		if isSeed {
-			regionIDCounter++
-			voxel.RegionID = regionIDCounter
-			voxel.Seed = true
-		} else if nearestNeighbor != nil {
-			voxel.RegionID = nearestNeighbor.RegionID
-		}
-
-		regionMap[voxel.RegionID] = append(regionMap[voxel.RegionID], [3]int{x, y, z})
-	}
-
-	regionCount := map[int]int{}
-	for x := 0; x < dimensions[0]; x++ {
-		for y := 0; y < dimensions[1]; y++ {
-			for z := 0; z < dimensions[2]; z++ {
-				if !voxelField[x][y][z].Filled {
-					continue
-				}
-
-				voxel := &voxelField[x][y][z]
-				regionCount[voxel.RegionID]++
-			}
-		}
-	}
-	fmt.Printf("generated %d regions\n", len(regionCount))
-	return regionMap
-}
-
-func getDebugNeighbors(x, y, z int, voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int) []*Voxel {
-	var neighbors []*Voxel
-
-	for _, dir := range neighborDirs {
-		if x+dir[0] < 0 || z+dir[1] < 0 || x+dir[0] >= dimensions[0] || z+dir[1] >= dimensions[2] {
-			neighbors = append(neighbors, nil)
-			continue
-		}
-
-		var neighbor *Voxel
-		reachNeighbor := &reachField[x+dir[0]][y][z+dir[1]]
-		if reachNeighbor.sourceVoxel != nil {
-			neighbor = reachNeighbor.sourceVoxel
-		} else if voxelField[x+dir[0]][y][z+dir[1]].Filled {
-			neighbor = &voxelField[x+dir[0]][y][z+dir[1]]
-		}
-
-		neighbors = append(neighbors, neighbor)
-	}
-
-	return neighbors
-}
-
 func getNeighbors(x, y, z int, voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int) []*Voxel {
 	var neighbors []*Voxel
 
@@ -520,142 +135,6 @@ func getNeighbors(x, y, z int, voxelField [][][]Voxel, reachField [][][]ReachInf
 	}
 
 	return neighbors
-}
-
-var blurWeights []float64 = []float64{
-	// 1, 2, 1,
-	// 2, 4, 2,
-	// 1, 2, 1,
-	1, 1, 1,
-	1, 1, 1,
-	1, 1, 1,
-}
-
-func blurDistanceField(voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int) {
-	blurredDistance := make([][][]float64, dimensions[0])
-	for i := range voxelField {
-		blurredDistance[i] = make([][]float64, dimensions[1])
-		for j := range voxelField[i] {
-			blurredDistance[i][j] = make([]float64, dimensions[2])
-		}
-	}
-
-	for x := 0; x < dimensions[0]; x++ {
-		for y := 0; y < dimensions[1]; y++ {
-			for z := 0; z < dimensions[2]; z++ {
-
-				var totalDistance float64
-				var totalWeight float64
-				for i, dir := range neighborDirs {
-					if x+dir[0] < 0 || z+dir[1] < 0 || x+dir[0] >= dimensions[0] || z+dir[1] >= dimensions[2] {
-						continue
-					}
-
-					var neighbor *Voxel
-					reachNeighbor := &reachField[x+dir[0]][y][z+dir[1]]
-					if reachNeighbor.sourceVoxel != nil {
-						neighbor = reachNeighbor.sourceVoxel
-					} else if voxelField[x+dir[0]][y][z+dir[1]].Filled {
-						neighbor = &voxelField[x+dir[0]][y][z+dir[1]]
-					}
-
-					if neighbor == nil {
-						continue
-					}
-
-					totalWeight += blurWeights[i]
-					totalDistance += blurWeights[i] * neighbor.DistanceField
-				}
-				totalWeight += blurWeights[4]
-				totalDistance += blurWeights[4] * voxelField[x][y][z].DistanceField
-
-				totalDistance /= float64(totalWeight)
-				blurredDistance[x][y][z] = totalDistance
-			}
-		}
-	}
-
-	for x := 0; x < dimensions[0]; x++ {
-		for y := 0; y < dimensions[1]; y++ {
-			for z := 0; z < dimensions[2]; z++ {
-				voxelField[x][y][z].DistanceField = blurredDistance[x][y][z]
-			}
-		}
-	}
-}
-
-func mergeRegions(voxelField [][][]Voxel, reachField [][][]ReachInfo, dimensions [3]int, regionMap map[int][][3]int) {
-	processedRegions := map[int]bool{}
-	regionConversion := map[int]int{}
-
-	var regionIDs []int
-	for id := range regionMap {
-		regionIDs = append(regionIDs, id)
-	}
-	sort.Ints(regionIDs)
-
-	for _, regionID := range regionIDs {
-		if _, ok := processedRegions[regionID]; ok {
-			continue
-		}
-
-		// the first coordinates always belongs to the seed voxel
-		coords := regionMap[regionID][0]
-		x := coords[0]
-		y := coords[1]
-		z := coords[2]
-
-		var search []*Voxel
-		search = append(search, &voxelField[x][y][z])
-		seen := map[[3]int]bool{}
-
-		for len(search) > 0 {
-			voxel := search[0]
-			search = search[1:]
-
-			if seen[[3]int{voxel.X, voxel.Y, voxel.Z}] {
-				continue
-			}
-
-			seen[[3]int{voxel.X, voxel.Y, voxel.Z}] = true
-			processedRegions[voxel.RegionID] = true
-			regionConversion[voxel.RegionID] = regionID
-
-			neighbors := getNeighbors(voxel.X, voxel.Y, voxel.Z, voxelField, reachField, dimensions)
-			for _, neighbor := range neighbors {
-				if !neighbor.Seed {
-					continue
-				}
-				if _, ok := processedRegions[neighbor.RegionID]; ok {
-					continue
-				}
-				if seen[[3]int{neighbor.X, neighbor.Y, neighbor.Z}] {
-					continue
-				}
-
-				search = append(search, neighbor)
-			}
-		}
-	}
-
-	for _, regionID := range regionIDs {
-		if _, ok := regionConversion[regionID]; !ok {
-			continue
-		}
-
-		nextRegionID := regionConversion[regionID]
-		if regionID == nextRegionID {
-			continue
-		}
-
-		coords := regionMap[regionID]
-		for _, coord := range coords {
-			x := coord[0]
-			y := coord[1]
-			z := coord[2]
-			voxelField[x][y][z].RegionID = nextRegionID
-		}
-	}
 }
 
 func (n *NavigationMesh) BakeNavMesh() {
